Better icon
Added Lanczos scaling in the form of an internal package rather than the hard-coded nearest-neighbor implementation. Makes it easier for me to reuse this code for other purposes, and opens the door for allowing the user to select the scaling algorithm.

I didn't use the golang.org/x/image/draw `CatmullRom` because I didn't want to introduce any dependency other than what is available in the standard library. This probably doesn't justify how long this took to get right.

The print package also introduces `HalfPrint` which produces better the previous method which was a rudimentary implementation of `BlockPrint`.